Add files via upload
Added a check to see if the Buy and Sell text files contained 12 months of pricing data. This will automatically stop the program from appending to the text files and give you
you're proper YTD return. You can delete the files and try to use the script to perform the Calendar Anomaly for any year, make sure you are diligent when testing the Calendar Anomaly.